change payload type to long and switch from ms to ns for execution time

+++ railo-java/railo-core/src/railo/runtime/cache/tag/CacheItem.java
@@ -6,7 +6,7 @@ public interface CacheItem {
 
 	public String getName();
 
-	public int getPayload();
+	public long getPayload();
 
 	public String getMeta();
